splitting reducers into sep files
