+++ src/store/storeTypes.ts
@@ -1,0 +1,4 @@
+export type CustomReducerAction<ReduxTypeGeneric, ReduxPayloadGeneric> = {
+    type: ReduxTypeGeneric;
+    payload: ReduxPayloadGeneric;
+};
